DemoRenderer: Don't try to create a texture of width zero. libGL might crash and definitely won't produce anything useful. This simple regression was caused by accidentally changing float corner_radius = 0.5f into an int in revision 1845. (LP: #1358210)

+++ examples/demo-shell/demo_renderer.cpp
@@ -77,7 +77,7 @@ GLuint generate_shadow_corner_texture(float opacity)
     return corner;
 }
 
-GLuint generate_frame_corner_texture(int corner_radius,
+GLuint generate_frame_corner_texture(float corner_radius,
                                      Color const& color,
                                      GLubyte highlight)
 {
